refactor(time-picker): ♻️  remove default value state & make utils
