+++ src/timepicker/TimePickerColumnState.ts
@@ -1,7 +1,9 @@
 import { useCompositeState } from "reakit";
 import { useControllableState } from "@chakra-ui/hooks";
 
-type ColumnType = "hour" | "minute" | "meridian";
+import { getSelectedValueFromDate, getSelectedDateFromValue } from "./__utils";
+
+export type ColumnType = "hour" | "minute" | "meridian";
 interface TimePickerColumnInitialProps {
   value?: Date;
   defaultValue?: Date;
@@ -45,41 +47,3 @@ export const useTimePickerColumnState = (
 export type TimePickerColumnStateReturn = ReturnType<
   typeof useTimePickerColumnState
 >;
-
-function getSelectedValueFromDate(date: Date, type: ColumnType) {
-  if (type === "minute") return date.getMinutes();
-  if (type === "meridian") return date.getHours() >= 12 ? 1 : 0;
-  return date.getHours() % 12 || 12;
-}
-
-function getSelectedDateFromValue(value: number, date: Date, type: ColumnType) {
-  if (type === "minute") {
-    return new Date(date.setMinutes(value));
-  }
-
-  if (type === "meridian") {
-    let hours = date.getHours() % 12;
-
-    if (value === 1) {
-      hours = hours + 12;
-    } else {
-      hours = hours % 12;
-    }
-
-    return new Date(date.setHours(hours));
-  }
-
-  if (date.getHours() >= 12) {
-    if (value !== 12) {
-      value = value + 12;
-    }
-  } else {
-    if (value !== 12) {
-      value = value % 12;
-    } else {
-      value = 0;
-    }
-  }
-
-  return new Date(date.setHours(value));
-}
